fix: rename entitiesCurrentPage().loading to isLoading
